test: split into separate test file
