+++ test/nuxt/client-only.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it, vi } from 'vitest'
+import { describe, expect, it } from 'vitest'
 import type { ComponentOptions } from 'vue'
 import { Suspense, defineComponent, h, toDisplayString, useAttrs } from 'vue'
 import { mountSuspended } from '@nuxt/test-utils/runtime'
@@ -88,29 +88,3 @@ function createWrappedClientPage () {
 
   return { resolve: resolve!, wrapper }
 }
-
-describe('app/compat', () => {
-  const Component = defineComponent({
-    setup () {
-      const visible = ref(false)
-      setInterval(() => {
-        visible.value = true
-      }, 1000)
-
-      return () => h('div', {}, visible.value ? h('span', { id: 'child' }) : {})
-    },
-  })
-  it('setInterval is not auto-imported', async () => {
-    vi.useFakeTimers()
-
-    const wrapper = mount(Component)
-
-    vi.advanceTimersByTime(1000)
-
-    await wrapper.vm.$nextTick()
-
-    expect(wrapper.find('#child').exists()).toBe(true)
-
-    vi.useRealTimers()
-  })
-})
